+++ setup.cfg
@@ -1,0 +1,28 @@
+[metadata]
+name = ppopt
+version = 0.1.0
+author = Kirill Sechkar
+author_email = kirill.sechkar18@imperial.ac.uk
+description = Saving pipette tips when distributing DNA parts in automated DNA assembly
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/KSechkar/pipette_opt
+project_urls =
+    Bug Tracker = https://github.com/KSechkar/pipette_opt/pipette_opt
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.8
+install_requires =
+    numpy>=1.17.1
+    gurobipy>=9.1.1
+    opentrons==3.21.1
+
+[options.packages.find]
+where = src
