+++ src/ppopt/__init__.py
@@ -1,0 +1,5 @@
+""""ppopt - algorithmic optimisation of pipette tip consumption in automated DNA assembly"""
+
+__version__ = '0.1.0'
+__author__ = 'Kirill Sechkar <kirill.sechkar18@imperial.ac.uk>'
+__all__ = ['statespace','lp','dp','auxil','basic','moclo','startstop','visualise']
