+++ src/ppopt/dp/__init__.py
@@ -1,0 +1,6 @@
+"""dp - Dynamic Programming optimisation method"""
+from .dp_method import dp_method
+
+__version__ = '0.1.0'
+__author__ = 'Kirill Sechkar <kirill.sechkar18@imperial.ac.uk>'
+__all__=['dp_method']
