+++ src/ppopt/lp/__init__.py
@@ -1,0 +1,6 @@
+"""lp - Linear Programming optimisation method"""
+from .lp_method import lp_method
+
+__version__ = '0.1.0'
+__author__ = 'Kirill Sechkar <kirill.sechkar18@imperial.ac.uk>'
+__all__=['lp_method','lp_solver','lp_reorder']
