+++ src/ppopt/statespace/__init__.py
@@ -1,0 +1,6 @@
+"""statespace - state-space graph searching optimisation methods"""
+from .statespace_methods import nns,greedy_tree
+
+__version__ = '0.1.0'
+__author__ = 'Kirill Sechkar <kirill.sechkar18@imperial.ac.uk>'
+__all__=['statespace_methods']
