+++ src/ppopt/test/__init__.py
@@ -1,0 +1,5 @@
+""""test - testing the algorithms' performance"""
+
+__version__ = '0.1.0'
+__author__ = 'Kirill Sechkar <kirill.sechkar18@imperial.ac.uk>'
+__all__ = []
